Separate getStoryID
