+++ lib/helpers/pivotaly.js
@@ -1,24 +1,6 @@
 const {common} = require("../commands/common")
-const {isStoryIDValid} = require("../validation/validators/isStoryIDValid")
 const isStoryAChore = require("../validation/validators/isStoryAChore")
-const {workspace} = require("vscode")
-
-const getStoryID = async (context, branch) => {
-  const delimiter = workspace.getConfiguration().get('pivotaly.branchDelimiter')
-  const splitBranch = branch.split(delimiter)
-  const postStoryID = splitBranch.pop()
-  const preStoryID = splitBranch[0]
-
-  const isPostStoryValid = await isStoryIDValid(context, postStoryID)
-  if(isPostStoryValid)
-    return postStoryID
   
-  const isPreStoryValid = await isStoryIDValid(context, preStoryID)
-  if(isPreStoryValid)
-    return preStoryID
-  return undefined
-}
-
 const getState = (context) => {
   const defaultState = JSON.stringify({branch: "", story: ""})
   return JSON.parse(context.workspaceState.get(common.globals.state, defaultState))
@@ -37,7 +19,6 @@ const setState = async (context, branch = "", story = "") => {
 }
 
 module.exports = {
-  getStoryID,
   getState,
   setState
 }
